moved getRedirectLink to utils.js
